+++ packages/appinio_video_player/example/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:appinio_video_player/appinio_video_player.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/foundation.dart';
 
 void main() {
   runApp(const MyApp());
@@ -32,79 +33,24 @@ class _MyHomePageState extends State<MyHomePage> {
   late VideoPlayerController _videoPlayerController,
       _videoPlayerController2,
       _videoPlayerController3;
+
   late CustomVideoPlayerController _customVideoPlayerController;
-  String videoUrlLandscape =
-      "https://flutter.github.io/assets-for-api-docs/assets/videos/bee.mp4";
-  String videoUrlPortrait =
-      'https://assets.mixkit.co/videos/preview/mixkit-a-girl-blowing-a-bubble-gum-at-an-amusement-park-1226-large.mp4';
-  String longVideo =
-      "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4";
-
-  String video720 =
-      "https://www.sample-videos.com/video123/mp4/720/big_buck_bunny_720p_10mb.mp4";
-
-  String video480 =
-      "https://www.sample-videos.com/video123/mp4/480/big_buck_bunny_480p_10mb.mp4";
-
-  String video240 =
-      "https://www.sample-videos.com/video123/mp4/240/big_buck_bunny_240p_10mb.mp4";
+  late CustomVideoPlayerWebController _customVideoPlayerWebController;
 
   final CustomVideoPlayerSettings _customVideoPlayerSettings =
-      const CustomVideoPlayerSettings(
-          //TODO: play around with these parameters
+      const CustomVideoPlayerSettings();
 
-          // autoFadeOutControls: false,
-          // controlBarAvailable: false,
-          // playbackSpeedButtonAvailable: false,
-          // controlBarMargin: const EdgeInsets.all(10),
-          // controlBarPadding: const EdgeInsets.all(0),
-          // showPlayButton: true,
-          // playButton: const Icon(
-          //   Icons.play_circle,
-          //   color: Colors.white,
-          // ),
-          // pauseButton: const Icon(
-          //   Icons.pause_circle,
-          //   color: Colors.white,
-          // ),
-          // enterFullscreenButton: const Icon(
-          //   Icons.fullscreen,
-          //   color: Colors.white,
-          // ),
-          // exitFullscreenButton: const Icon(
-          //   Icons.fullscreen_exit,
-          //   color: Colors.white,
-          // ),
-          // controlBarDecoration: BoxDecoration(
-          //   borderRadius: BorderRadius.circular(50),
-          //   color: Colors.blue,
-          // ),
-          // showFullscreenButton: false,
-          // showDurationPlayed: false,
-          // showDurationRemaining: false,
-          // enterFullscreenOnStart: true,
-          // exitFullscreenOnEnd: true,
-          // durationRemainingTextStyle: const TextStyle(color: Colors.red),
-          // durationPlayedTextStyle: const TextStyle(color: Colors.green),
-          // systemUIModeAfterFullscreen: SystemUiMode.leanBack,
-          // systemUIModeInsideFullscreen: SystemUiMode.edgeToEdge,
-          // customVideoPlayerProgressBarSettings: CustomVideoPlayerProgressBarSettings(
-          //   reachableProgressBarPadding: EdgeInsets.all(10),
-          //   progressBarHeight: 10,
-          //   progressBarBorderRadius: 30,
-          //   bufferedColor: Colors.red,
-          //   progressColor: Colors.green,
-          //   backgroundColor: Colors.purple,
-          //   allowScrubbing: false,
-          //   showProgressBar: false,
-          // ),
-          );
+  final CustomVideoPlayerWebSettings _customVideoPlayerWebSettings =
+      CustomVideoPlayerWebSettings(
+    src: longVideo,
+  );
 
   @override
   void initState() {
     super.initState();
+
     _videoPlayerController = VideoPlayerController.network(
-      video720,
+      longVideo,
     )..initialize().then((value) => setState(() {}));
     _videoPlayerController2 = VideoPlayerController.network(video240);
     _videoPlayerController3 = VideoPlayerController.network(video480);
@@ -117,6 +63,10 @@ class _MyHomePageState extends State<MyHomePage> {
         "480p": _videoPlayerController3,
         "720p": _videoPlayerController,
       },
+    );
+
+    _customVideoPlayerWebController = CustomVideoPlayerWebController(
+      webVideoPlayerSettings: _customVideoPlayerWebSettings,
     );
   }
 
@@ -135,15 +85,26 @@ class _MyHomePageState extends State<MyHomePage> {
       child: SafeArea(
         child: ListView(
           children: [
-            CustomVideoPlayer(
-              customVideoPlayerController: _customVideoPlayerController,
-            ),
+            kIsWeb
+                ? Expanded(
+                    child: CustomVideoPlayerWeb(
+                      customVideoPlayerWebController:
+                          _customVideoPlayerWebController,
+                    ),
+                  )
+                : CustomVideoPlayer(
+                    customVideoPlayerController: _customVideoPlayerController,
+                  ),
             CupertinoButton(
               child: const Text("Play Fullscreen"),
               onPressed: () {
-                // example on how to set fullscreen yourself
-                // _customVideoPlayerController.setFullscreen(true);
-                _customVideoPlayerController.videoPlayerController.play();
+                if (kIsWeb) {
+                  _customVideoPlayerWebController.setFullscreen(true);
+                  _customVideoPlayerWebController.play();
+                } else {
+                  _customVideoPlayerController.setFullscreen(true);
+                  _customVideoPlayerController.videoPlayerController.play();
+                }
               },
             ),
           ],
@@ -152,3 +113,19 @@ class _MyHomePageState extends State<MyHomePage> {
     );
   }
 }
+
+String videoUrlLandscape =
+    "https://flutter.github.io/assets-for-api-docs/assets/videos/bee.mp4";
+String videoUrlPortrait =
+    'https://assets.mixkit.co/videos/preview/mixkit-a-girl-blowing-a-bubble-gum-at-an-amusement-park-1226-large.mp4';
+String longVideo =
+    "https://commondatastorage.googleapis.com/gtv-videos-bucket/sample/BigBuckBunny.mp4";
+
+String video720 =
+    "https://www.sample-videos.com/video123/mp4/720/big_buck_bunny_720p_10mb.mp4";
+
+String video480 =
+    "https://www.sample-videos.com/video123/mp4/480/big_buck_bunny_480p_10mb.mp4";
+
+String video240 =
+    "https://www.sample-videos.com/video123/mp4/240/big_buck_bunny_240p_10mb.mp4";