+++ packages/appinio_video_player/lib/src/web_video_player/conditional_native_web_video_player/conditional_native_web_video_player_base.dart
@@ -1,0 +1,4 @@
+import 'conditional_native_web_video_player.dart';
+
+ConditionalNativeWebVideoPlayer getSomething() =>
+    throw UnsupportedError("Cannot create NativeWebVideoPlayer");