+++ packages/appinio_video_player/lib/src/web_video_player/custom_video_player_web.dart
@@ -1,0 +1,16 @@
+import 'package:appinio_video_player/appinio_video_player.dart';
+import 'package:flutter/material.dart';
+
+class CustomVideoPlayerWeb extends StatelessWidget {
+  final CustomVideoPlayerWebController customVideoPlayerWebController;
+  const CustomVideoPlayerWeb({
+    Key? key,
+    required this.customVideoPlayerWebController,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ConditionalNativeWebVideoPlayer()
+        .getVideoWidget(customVideoPlayerWebController);
+  }
+}